Use existing kubeconfig from workspace

+++ entrypoint.sh
@@ -4,7 +4,12 @@ set -e
 
 echo -e "\033[36mSetting up kubectl configuration\033[0m"
 mkdir ~/.kube/ || true
-echo "${INPUT_KUBECONFIG}" > ~/.kube/config
+if [ -f "~/.kube/config" ]; then
+  echo -e "\033[36mExisting kubeconfig found, using that and ignoring input\033[0m"
+else
+  echo -e "\033[36mUsing kubeconfig from input\033[0m"
+  echo "${INPUT_KUBECONFIG}" > ~/.kube/config
+fi
 
 echo -e "\033[36mPreparing helm execution\033[0m"
 echo "${INPUT_EXEC}" > run.sh
